refactor: abstracts 폴더 생성
