+++ src/abstracts/constants/index.ts
@@ -1,0 +1,1 @@
+export const CARD_LIST_STORAGE_KEY = 'cardList';
