+++ src/components/box/inputSection/ExpireDateInput.tsx
@@ -1,6 +1,6 @@
 import InputSectionTemplate from '../../template/InputSectionTemplate';
 import InputBox, { InputType } from '../../common/InputBox';
-import { InputStateProps } from '../../../types';
+import { InputStateProps } from '../../../abstracts/types';
 
 const ExpireDateInput = (props: InputStateProps) => {
   const inputs: InputType[] = [
